feat(moq-transcode)!: resolve the rung ladder into a canonical order
`Config::rungs` was a plain `Vec<Rung>` with no ordering contract: it was
filtered against the source in configuration order, and duplicate heights were
silently dropped. Nothing downstream could say what "the next lower rendition"
meant, which is the input the congestion-aware ladder work needs.

`Ladder` now owns that contract. `Ladder::new` takes the rungs in any order,
rounds each height down to even, and ranks them by configured maximum bitrate,
lowest first. A set of rungs that cannot be ranked is refused instead of
guessed: two rungs sharing a maximum, a rung that costs more without being
taller (equal heights included, since they would collide on one track name),
and a rung with no picture or no bitrate. `Config::rungs` holds a `Ladder`
rather than a `Vec`, so an unorderable ladder cannot reach the transcoder.

Resolving against the source drops rungs but never reorders them, so the
resolved ladder stays ascending. The CLI builds the ladder before it dials, so
an operator typo is reported immediately rather than after the relay and the
source announcement.

Also fixes an unrelated dead-code break on `dev`: `DmaBufExport::inner` is only
read by the render feature, so a moq-cli + moq-transcode selection failed
`just check` before this branch changed anything.
