+++ rs/CLAUDE.md
@@ -28,7 +28,7 @@ Layered roughly transport -> container/format -> media -> apps/bindings.
 - `moq-mux` (lib): the conversion layer. File/stream formats (`container/`: fmp4, flv, mkv, ts, loc) and codec parsers (`codec/`: h264, h265, av1, vp8/9, opus, aac, ...) <-> hang broadcasts. `Container` trait + generic `Producer<C>`/`Consumer<C>`. Dual catalog (`catalog::hang`, `catalog::msf`).
 - `moq-audio` (lib): native PCM <-> Opus/PCM (`unsafe-libopus`), plus AAC-LC decode (`symphonia-codec-aac`, the default-on `aac` feature) for broadcasts that arrived through a gateway. Shaped like `moq-video`: `capture::Config`, `encode::{Encoder, Producer, Publication, PublicationOptions, Driver, publish_capture}`, `decode::{Consumer, Decoder}`, plus root `Error`/`Format`/`Frame`. `Publication` controls a stable capture-backed track (start, stop, input replacement, device/failure/level state) while `Driver` owns capture and encoding. Playback is the extra role module `moq-video` has no counterpart for: `playback::{Engine, Config, Sink, Control, Input, Device, devices}`, where one `Engine` owns the output device and mixes up to 64 `Sink`s into it on a driver thread. `aec::{Canceller, Config}` closes the loop between the two: `Engine::canceller` taps the post-mix signal and `capture::Config::aec` subtracts it from the microphone (`sonora`, a pure-Rust WebRTC APM port), which is what a call on a laptop needs to not send itself back. Optional `capture` feature (cpal microphone, macOS system audio), `playback` feature (cpal output, `fixed-resample` ring buffers), and `aec` feature (implies both).
 - `moq-video` (lib): native video capture, H.264/H.265 encode, and decode; no ffmpeg. Hardware backends (VideoToolbox / Media Foundation / NVENC / VAAPI / NVDEC) with openh264 as the software H.264 fallback; NVDEC frames stay in CUDA memory and feed NVENC zero-copy. `capture::Config`, `encode::{Encoder, Producer, publish_capture}`, `decode::{Consumer, Decoder}`, `convert::{Config, Rgba}`, root `Error`/`Size`.
-- `moq-transcode` (lib): just-in-time live transcoding of hang broadcasts. `Transcoder::new(source, output, config)` registers the output tracks synchronously (announce `output` after it) and `run` drives it: a derivative catalog (ladder rungs + relative refs to the source) plus each rung encoded only while subscribed/fetched, via `moq-video`. The free `run(source, output, config)` is the shorthand. `Transcoder::active` hands out `active::Consumer` cursors over which renditions are encoding, each `active::Rendition` counting the frames and bytes it produced for a caller that bills. Live rungs share one decode per source (the `feed` module); output groups mirror source group sequences 1:1. Also a moq-cli verb (`moq ... transcode`, feature-gated).
+- `moq-transcode` (lib): just-in-time live transcoding of hang broadcasts. `Transcoder::new(source, output, config)` registers the output tracks synchronously (announce `output` after it) and `run` drives it: a derivative catalog (ladder rungs + relative refs to the source) plus each rung encoded only while subscribed/fetched, via `moq-video`. The free `run(source, output, config)` is the shorthand. `Transcoder::active` hands out `active::Consumer` cursors over which renditions are encoding, each `active::Rendition` counting the frames and bytes it produced for a caller that bills. Live rungs share one decode per source (the `feed` module); output groups mirror source group sequences 1:1. `Config::rungs` is a `Ladder`: `Ladder::new` takes the rungs in any order and ranks them by configured maximum bitrate, lowest first, refusing a ladder that cannot be ranked (a shared bitrate, or a rung that costs more without being taller) so "the next lower rendition" always names one. Also a moq-cli verb (`moq ... transcode`, feature-gated).
 
 **Apps / binaries**
